CI: don't run commands through pipenv
Since #62, the dependencies have been installed to the system
interpreter. Therefore, it's not necessary to run through pipenv to
ensure the commands run through an activated virtual environment.

This should also fix pipenv run creating a virtual environment. It seems
it cannot tell when it should be using the system interpreter. It
probably wasn't designed for that anyway i.e. the intent was to run
commands directly in such case, which is what this PR will do.