CI: output flake8 to stdout (resolve #37)
The Azure website has proven to not be reliable for displaying the
JUnit output. Furthermore, some may simply prefer the format of the
output in the terminal over the JUnit representation on the Azure site.

Nevertheless, the JUnit output isn't that bad (when there's actually
a lint error) so it will remain for now. It also provides historical
statistics on occurrences of errors, which is kind of cool, I guess...