CI: fix can_pull causing script to exit with code 1
Error handling is performed inside can_pull so the callers of the
function don't always check its exit code. Because set -e present, if
can_pull returns 1, bash would consider that function a failed call and
thus exit the entire script with code 1. That, in turn, would cause the
CI job to fail.