Migrate build pipeline to Github Actions
I've migrated the build pipeline to GitHub Actions and changed the
container registry to GitHub Container Registry. In the process, I've
made some changes to our docker setup and caching:

- We are now using a single multi-stage Dockerfile

Instead of three separate dockerfiles, we are now using a
single multi-stage Dockerfile that can be used to build the three images
we want using build targets.

In part, this is because we're now using the docker buildx build action
currently recommended by docker. This new engine runs in a sandboxed
mode, meaning that while it can export built images to `docker` running
in the host, it cannot import local images from it to base builds on.

- Docker builds are now cached within GitHub Actions

The builds are now cached using the GitHub Actions cache of the build
cache directory. The cache keys try to match a cache generated by a
build that matches the current build as closely as possible. In case of
a cache miss, we fall back to caching from the latest image pushed to
the container repository.

- The `base` and `venv` images now have an inline cache manifest

In order to fall back intelligently to caching from the repository, the
final build and push action for the `base` and `venv` images includes an
"inline" cache manifest. This means that the build process can inspect,
without pulling, if it makes sense to pull layers to speed up the build.

The other options, pushing a cache manifest separately (not inline), is
currently not supported by GHCR.

The custom caching script has been removed.

- Linting errors are now added as GitHub Actions annotations

Just like for some of our other pipelines, linting now generates
annotations if linting errors are observed.

- Coverage is pushed to coveralls.io

A coverage summary is now pushed to coveralls.io. Each CI run will get a
unique job that's linked in the CI output. If the run is attached to a
PR, coveralls.io will automatically add a check link with the coverage
result to the PR as well.

- The README.md, Pipfile, docker-compose, and scripts have been updated

As we now need to pull from and link to the GHCR, I've updated the other
files to reflect these changes, including Pipfile run commands. I've
also changed the CI badge and added a coveralls.io badge.