Fix caching of exposed packages
The problem with --ignore-installed is that it also ignores packages in
the target site, therefore re-installing all packages and making caching
futile.

Install snekbox's dependencies to the default user site. During
installation of the exposed packages, switch the user site to a custom
one to ensure pip will not skip packages due to seeing them as already
installed as one of snekbox's dependencies.

If pip made the --root option ignore packages outside the root, then
using --root would be the best solution. There is a 5+ year old issue
open about that.