CI: run linters outside the container
Remove the dependency on the container so the lint job can run in
parallel with the build job. More time has to be spent installing
Python dependencies, but this is made up for by not having to download
and load the image artefact in addition to not having to wait for the
build job.