CI: clean up Docker stuff in self-hosted runner
Use docker-compose run instead of docker-compose up. This is more
appropriate since the container is only needed for one command. The
latter was actually starting the whole snekbox server. Furthermore,
the former has the --rm option to remove the container when the command
finishes.

As an extra precaution, use docker-compose down in the self-hosted
runner to also remove images, volumes, networks, and any other
containers that were somehow missed. Removing images will also prevent
the disk usage from building up. This is not necessary for the GH-hosted
runner since a new VM is used for each run.