Make Sentry SDK and gunicorn optional dependencies
Falcon provides a WSGI app which can be used by any server, not just
gunicorn. Thus, make gunicorn optional in case the user wants to use
a different server. There shouldn't be any import errors since the class
is now in an isolated module. The only time that module is imported is
when gunicorn loads its config.

Sentry is there for Python Discord mainly, so this dependency shouldn't
be imposed on others.