Update nsjail_args and py_args to store_const instead of nargs='?'
When using nargs='?' the nsjail_args argument was sometimes consuming 1 of the arguments passed to the module.
By setting these two arguments to store_const actions instead, this is no longer the case, as they will always use the constant specified now.
This is more appropriate for their actual use case of being purely for documentation.

Co-authored-by: GDWR <gregory.dwr@gmail.com>