Increase thread limit to 15
Changes to multiprocessing in Python 3.14 make it more likely that the 5
PID limit is hit even with non-complicated uses of multiprocessing.

We have enough compute to allocate more PIDs and safely know this will
not affect the operation of other services (we have since migrated our
databases and heavier processing applications onto other hosts).